+++ metadata.rb
@@ -1,10 +1,10 @@
-name             'autofs-nativex'
-maintainer       'NativeX'
-maintainer_email 'adrian.herrera@nativex.com'
+name             'autofs-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures autofs-nativex'
+description      'Installs/Configures autofs-simplyadrian'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '1.0.0'
 
-depends 'ohai-nativex'
+depends 'ohai-simplyadrian'
 depends 'chef-sugar'